// src/app/page/master/type-profile/type-profile.interface.ts
import { EventEmitter } from '@angular/core';
export interface TypeProfile {

  typeProfile?: any;
  kode?: any;
  id?: any;
  kodeExternal?: string;
  namaExternal?: string;
  reportDisplay?: string;
  statusEnabled?: true;
  version?: any;
  Version?: any;
  kdTypeProfileHead?: any;
  kodeTypePegawai?: any;
  label?: any;
  value?: any;


}